# Conflicts:
#	angel-coreservice/angel-admin-api/src/main/java/com/pingan/angel/admin/api/mysql/DeviceEntity.java

+++ angel-coreservice/angel-admin-api/src/main/java/com/pingan/angel/admin/api/mysql/DeviceStatusEntity.java
@@ -21,8 +21,8 @@ public class DeviceStatusEntity extends BaseEntity{
     private double totalCleanWater;//总纯水量
     private double totalUsedWater;//历史总用水量
     private Date reportTime;//最后上报时间
-    private int outTds;//出水TDS
-    private int inTds;//进水TDS
+    private double outTds;//出水TDS
+    private double inTds;//进水TDS
     private String isStartUp;//设备开机状态
     private Date lastControlTime;//最后一次控制时间
     private String configUnit;//上报单位，默认h ，可用值h d 小时或者天,开发需要判断并换上成小时处理
